refactor(plugin): extract limited/{texture,buffer,texture_registration}.rs (#1092)
Continue splitting gpu_context/mod.rs's LimitedAccess vtable section.

* limited/texture.rs (~400 LoC) — Texture Arc-handle lifecycle +
  Texture::native_handle DMA-BUF FD export (Phase F #957) +
  PooledTextureHandle drop + texture-cache method dispatch
  (register, update layout, acquire, resolve, unregister).

* limited/buffer.rs (~420 LoC) — Linux-only StorageBuffer /
  UniformBuffer / VertexBuffer / IndexBuffer Arc-handle clone/drop
  pair + acquire_*_buffer method dispatch (v5). Per-type callbacks
  are individually addressable in the vtable so future divergence
  doesn't force a re-version; on non-Linux the buffer types don't
  exist and the slots compile to error returns for ABI stability.

* limited/texture_registration.rs (~225 LoC) — Arc<TextureRegistration>
  clone/drop + per-method dispatch (texture, current_layout,
  update_layout, resolve_by_surface_id). v6 wire-format anchor for
  docs/architecture/texture-registration.md.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored across all three commits.
