+++ libs/streamlib-engine/src/core/plugin/host_services/gpu_context/mod.rs
@@ -30,9 +30,22 @@ mod scope_token;
 mod shared;
 
 use limited::{
-    host_gpu_lim_clear_video_source_timeline_semaphore, host_gpu_lim_escalate_begin,
+    host_gpu_lim_acquire_index_buffer, host_gpu_lim_acquire_storage_buffer,
+    host_gpu_lim_acquire_texture, host_gpu_lim_acquire_uniform_buffer,
+    host_gpu_lim_acquire_vertex_buffer, host_gpu_lim_clear_video_source_timeline_semaphore,
+    host_gpu_lim_clone_index_buffer, host_gpu_lim_clone_storage_buffer, host_gpu_lim_clone_texture,
+    host_gpu_lim_clone_texture_registration, host_gpu_lim_clone_uniform_buffer,
+    host_gpu_lim_clone_vertex_buffer, host_gpu_lim_drop_index_buffer,
+    host_gpu_lim_drop_pooled_texture_handle, host_gpu_lim_drop_storage_buffer,
+    host_gpu_lim_drop_texture, host_gpu_lim_drop_texture_registration,
+    host_gpu_lim_drop_uniform_buffer, host_gpu_lim_drop_vertex_buffer, host_gpu_lim_escalate_begin,
     host_gpu_lim_escalate_end, host_gpu_lim_host_video_source_timeline_arc,
-    host_gpu_lim_set_video_source_timeline_semaphore, host_gpu_lim_wait_timeline_semaphore,
+    host_gpu_lim_register_texture, host_gpu_lim_resolve_texture_by_surface_id,
+    host_gpu_lim_resolve_texture_registration_by_surface_id,
+    host_gpu_lim_set_video_source_timeline_semaphore, host_gpu_lim_texture_native_dma_buf_fd,
+    host_gpu_lim_texture_registration_current_layout, host_gpu_lim_texture_registration_texture,
+    host_gpu_lim_texture_registration_update_layout, host_gpu_lim_unregister_texture,
+    host_gpu_lim_update_texture_registration_layout, host_gpu_lim_wait_timeline_semaphore,
 };
 use scope_token::with_full_scope_or_err;
 use shared::{handle_as_gpu_context, pixel_format_from_raw};
@@ -204,980 +217,6 @@ unsafe extern "C" fn host_gpu_lim_plane_size_pixel_buffer(
         0,
     )
 }
-
-// -------------------------------------------------------------------------
-// Texture Arc-handle lifecycle
-// -------------------------------------------------------------------------
-
-unsafe extern "C" fn host_gpu_lim_clone_texture(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_clone_texture",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: `handle` is a `*const c_void` cast of
-            // `Arc::into_raw(Arc<TextureInner>)` produced by host
-            // code (see `Texture::from_arc_into_raw`).
-            unsafe {
-                Arc::increment_strong_count(
-                    handle as *const crate::core::rhi::texture::TextureInner,
-                );
-            }
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_drop_texture(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_drop_texture",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: matched with the `Arc::into_raw` in
-            // `Texture::from_arc_into_raw` and any prior
-            // `clone_texture` bumps.
-            unsafe {
-                Arc::decrement_strong_count(
-                    handle as *const crate::core::rhi::texture::TextureInner,
-                );
-            }
-        },
-        (),
-    )
-}
-
-// -------------------------------------------------------------------------
-// Texture::native_handle DMA-BUF FD export (Phase F, #957)
-// -------------------------------------------------------------------------
-
-unsafe extern "C" fn host_gpu_lim_texture_native_dma_buf_fd(
-    texture_handle: *const c_void,
-) -> i64 {
-    run_host_extern_c(
-        "host_gpu_lim_texture_native_dma_buf_fd",
-        || {
-            if texture_handle.is_null() {
-                return -1;
-            }
-            #[cfg(target_os = "linux")]
-            {
-                // SAFETY: `texture_handle` is the
-                // `Arc::into_raw(Arc<TextureInner>)` pointer carried as the
-                // cdylib-side `Texture::handle` field. Borrowing as
-                // `&TextureInner` does not touch the refcount — the
-                // caller's `Texture` keeps the Arc alive for the duration
-                // of this dispatch.
-                let inner = unsafe {
-                    &*(texture_handle as *const crate::core::rhi::texture::TextureInner)
-                };
-                match inner.inner.export_dma_buf_fd() {
-                    Ok(fd) => i64::from(fd),
-                    Err(_) => -1,
-                }
-            }
-            #[cfg(not(target_os = "linux"))]
-            {
-                // DMA-BUF is a Linux concept. macOS / Windows native
-                // handles are deferred until those cdylib adapter paths
-                // resume (see #908's AI Agent Notes).
-                let _ = texture_handle;
-                -1
-            }
-        },
-        -1,
-    )
-}
-
-// -------------------------------------------------------------------------
-// PooledTextureHandle lifecycle — drop-only (v4)
-// -------------------------------------------------------------------------
-
-unsafe extern "C" fn host_gpu_lim_drop_pooled_texture_handle(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_drop_pooled_texture_handle",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: matched with `Box::into_raw(Box<...>)` in
-            // `PooledTextureHandle::from_parts`. Reclaiming via
-            // `Box::from_raw` runs `Drop for PooledTextureHandleInner`
-            // which releases the pool slot exactly once.
-            unsafe {
-                let _ = Box::from_raw(
-                    handle as *mut crate::core::context::texture_pool::PooledTextureHandleInner,
-                );
-            }
-        },
-        (),
-    )
-}
-
-// -------------------------------------------------------------------------
-// Method dispatch — Texture-related (v4)
-// -------------------------------------------------------------------------
-
-unsafe extern "C" fn host_gpu_lim_register_texture(
-    handle: *const c_void,
-    id_ptr: *const u8,
-    id_len: usize,
-    texture_handle: *const c_void,
-    initial_layout_raw: i32,
-) {
-    run_host_extern_c(
-        "host_gpu_lim_register_texture",
-        || {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                return;
-            };
-            if texture_handle.is_null() {
-                return;
-            }
-            // SAFETY: `texture_handle` is `Arc::into_raw(Arc<TextureInner>)`-shaped.
-            // Bump the refcount so we can hand the cache its own owned
-            // Arc; the caller's Texture continues to own its own.
-            unsafe {
-                Arc::increment_strong_count(
-                    texture_handle as *const crate::core::rhi::texture::TextureInner,
-                );
-            }
-            // SAFETY: same shape as above; from_raw + the bump above
-            // gives us a fresh Arc with the right refcount.
-            let texture_arc = unsafe {
-                Arc::from_raw(
-                    texture_handle as *const crate::core::rhi::texture::TextureInner,
-                )
-            };
-            let inner_ref = &*texture_arc;
-            let width = inner_ref.width();
-            let height = inner_ref.height();
-            let format = inner_ref.format();
-            // Re-wrap into a Texture via the host's from_arc_into_raw
-            // helper — leaks the Arc back into the texture cache shape.
-            let texture =
-                crate::core::rhi::texture::Texture::from_arc_into_raw(
-                    texture_arc, width, height, format,
-                );
-            let id_bytes = unsafe { slice_from_raw(id_ptr, id_len) };
-            let id_str = match std::str::from_utf8(id_bytes) {
-                Ok(s) => s,
-                Err(_) => return,
-            };
-            #[cfg(target_os = "linux")]
-            {
-                let layout = streamlib_consumer_rhi::VulkanLayout(initial_layout_raw);
-                gpu.register_texture_with_layout(id_str, texture, layout);
-            }
-            #[cfg(not(target_os = "linux"))]
-            {
-                let _ = initial_layout_raw;
-                gpu.register_texture(id_str, texture);
-            }
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_update_texture_registration_layout(
-    handle: *const c_void,
-    id_ptr: *const u8,
-    id_len: usize,
-    layout_raw: i32,
-) {
-    run_host_extern_c(
-        "host_gpu_lim_update_texture_registration_layout",
-        || {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                return;
-            };
-            let id_bytes = unsafe { slice_from_raw(id_ptr, id_len) };
-            let id_str = match std::str::from_utf8(id_bytes) {
-                Ok(s) => s,
-                Err(_) => return,
-            };
-            #[cfg(target_os = "linux")]
-            {
-                let layout = streamlib_consumer_rhi::VulkanLayout(layout_raw);
-                gpu.update_texture_registration_layout(id_str, layout);
-            }
-            #[cfg(not(target_os = "linux"))]
-            {
-                let _ = (id_str, layout_raw);
-            }
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_acquire_texture(
-    handle: *const c_void,
-    width: u32,
-    height: u32,
-    format_raw: u32,
-    usage_bits: u32,
-    out_pooled_handle: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_acquire_texture",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                write_err("acquire_texture: null gpu handle", err_buf, err_buf_cap, err_len);
-                return 1;
-            };
-            if out_pooled_handle.is_null() {
-                write_err("acquire_texture: null out_pooled_handle", err_buf, err_buf_cap, err_len);
-                return 1;
-            }
-            let format = match format_raw {
-                0 => streamlib_consumer_rhi::TextureFormat::Rgba8Unorm,
-                1 => streamlib_consumer_rhi::TextureFormat::Rgba8UnormSrgb,
-                2 => streamlib_consumer_rhi::TextureFormat::Bgra8Unorm,
-                3 => streamlib_consumer_rhi::TextureFormat::Bgra8UnormSrgb,
-                4 => streamlib_consumer_rhi::TextureFormat::Rgba16Float,
-                5 => streamlib_consumer_rhi::TextureFormat::Rgba32Float,
-                6 => streamlib_consumer_rhi::TextureFormat::Nv12,
-                _ => {
-                    let msg = format!("acquire_texture: invalid format_raw {}", format_raw);
-                    write_err(&msg, err_buf, err_buf_cap, err_len);
-                    return 1;
-                }
-            };
-            let usage =
-                streamlib_consumer_rhi::TextureUsages::from_bits_truncate(usage_bits);
-            let desc = crate::core::context::TexturePoolDescriptor {
-                width,
-                height,
-                format,
-                usage,
-                label: None,
-            };
-            match gpu.acquire_texture(&desc) {
-                Ok(pooled) => {
-                    // Move the host-built PooledTextureHandle into the
-                    // caller's out-slot. The caller (cdylib) owns it
-                    // after this — its Drop runs `drop_pooled_texture_handle`.
-                    unsafe {
-                        std::ptr::write(
-                            out_pooled_handle
-                                as *mut crate::core::context::PooledTextureHandle,
-                            pooled,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    let msg = format!("{}", e);
-                    write_err(&msg, err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_resolve_texture_by_surface_id(
-    handle: *const c_void,
-    surface_id_ptr: *const u8,
-    surface_id_len: usize,
-    has_layout: i32,
-    layout_raw: i32,
-    width: u32,
-    height: u32,
-    out_texture: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_resolve_texture_by_surface_id",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                write_err(
-                    "resolve_texture_by_surface_id: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if out_texture.is_null() {
-                write_err(
-                    "resolve_texture_by_surface_id: null out_texture",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            let id_bytes = unsafe { slice_from_raw(surface_id_ptr, surface_id_len) };
-            let id_str = match std::str::from_utf8(id_bytes) {
-                Ok(s) => s,
-                Err(_) => {
-                    write_err(
-                        "resolve_texture_by_surface_id: surface_id not valid UTF-8",
-                        err_buf,
-                        err_buf_cap,
-                        err_len,
-                    );
-                    return 1;
-                }
-            };
-            let texture_layout = if has_layout != 0 {
-                Some(layout_raw)
-            } else {
-                None
-            };
-            match gpu.resolve_texture_by_surface_id(id_str, texture_layout, width, height) {
-                Ok(texture) => {
-                    // Hand the texture to the caller's out-slot. The
-                    // caller (cdylib) owns it after this — its Drop
-                    // runs `drop_texture`.
-                    unsafe {
-                        std::ptr::write(
-                            out_texture as *mut crate::core::rhi::Texture,
-                            texture,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    let msg = format!("{}", e);
-                    write_err(&msg, err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_unregister_texture(
-    handle: *const c_void,
-    id_ptr: *const u8,
-    id_len: usize,
-) {
-    run_host_extern_c(
-        "host_gpu_lim_unregister_texture",
-        || {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                return;
-            };
-            let id_bytes = unsafe { slice_from_raw(id_ptr, id_len) };
-            let id_str = match std::str::from_utf8(id_bytes) {
-                Ok(s) => s,
-                Err(_) => return,
-            };
-            gpu.unregister_texture(id_str);
-        },
-        (),
-    )
-}
-
-// -------------------------------------------------------------------------
-// Linux-only buffer Arc-handle lifecycle
-// -------------------------------------------------------------------------
-//
-// All 4 buffer types (`StorageBuffer`, `UniformBuffer`, `VertexBuffer`,
-// `IndexBuffer`) wrap `Arc<HostVulkanBuffer>` under the hood. The per-
-// type callbacks are individually addressable in the vtable (so future
-// per-type divergence doesn't force a re-version) but share the same
-// host-side bookkeeping today. On non-Linux hosts the buffer types
-// don't exist, so the callbacks compile to no-ops / error returns —
-// the vtable slot is unconditional for ABI stability.
-
-#[cfg(target_os = "linux")]
-unsafe extern "C" fn host_gpu_lim_clone_host_vulkan_buffer_arc(handle: *const c_void) {
-    if handle.is_null() {
-        return;
-    }
-    // SAFETY: `handle` is `Arc::into_raw(Arc<HostVulkanBuffer>)`-shaped
-    // (see each buffer type's `from_arc_into_raw` constructor).
-    unsafe {
-        Arc::increment_strong_count(handle as *const crate::vulkan::rhi::HostVulkanBuffer);
-    }
-}
-
-#[cfg(target_os = "linux")]
-unsafe extern "C" fn host_gpu_lim_drop_host_vulkan_buffer_arc(handle: *const c_void) {
-    if handle.is_null() {
-        return;
-    }
-    // SAFETY: matched with the `Arc::into_raw` in each buffer type's
-    // `from_arc_into_raw` constructor.
-    unsafe {
-        Arc::decrement_strong_count(handle as *const crate::vulkan::rhi::HostVulkanBuffer);
-    }
-}
-
-#[cfg(not(target_os = "linux"))]
-unsafe extern "C" fn host_gpu_lim_clone_host_vulkan_buffer_arc(_handle: *const c_void) {
-    // Buffer types only exist on Linux; this callback is unreachable
-    // on other platforms. Defensive no-op.
-}
-
-#[cfg(not(target_os = "linux"))]
-unsafe extern "C" fn host_gpu_lim_drop_host_vulkan_buffer_arc(_handle: *const c_void) {
-    // Buffer types only exist on Linux; defensive no-op.
-}
-
-// Per-type wrappers. Each just delegates to the shared
-// `host_vulkan_buffer_arc` pair today but lives in the vtable as a
-// dedicated slot, so a future per-type divergence (e.g. UniformBuffer
-// growing a per-type cached field that needs its own clone semantics)
-// only edits the wrapper without touching the vtable surface.
-
-unsafe extern "C" fn host_gpu_lim_clone_storage_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_clone_storage_buffer",
-        || unsafe { host_gpu_lim_clone_host_vulkan_buffer_arc(handle) },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_drop_storage_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_drop_storage_buffer",
-        || unsafe { host_gpu_lim_drop_host_vulkan_buffer_arc(handle) },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_clone_uniform_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_clone_uniform_buffer",
-        || unsafe { host_gpu_lim_clone_host_vulkan_buffer_arc(handle) },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_drop_uniform_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_drop_uniform_buffer",
-        || unsafe { host_gpu_lim_drop_host_vulkan_buffer_arc(handle) },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_clone_vertex_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_clone_vertex_buffer",
-        || unsafe { host_gpu_lim_clone_host_vulkan_buffer_arc(handle) },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_drop_vertex_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_drop_vertex_buffer",
-        || unsafe { host_gpu_lim_drop_host_vulkan_buffer_arc(handle) },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_clone_index_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_clone_index_buffer",
-        || unsafe { host_gpu_lim_clone_host_vulkan_buffer_arc(handle) },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_drop_index_buffer(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_drop_index_buffer",
-        || unsafe { host_gpu_lim_drop_host_vulkan_buffer_arc(handle) },
-        (),
-    )
-}
-
-// -------------------------------------------------------------------------
-// Linux-only acquire_*_buffer method dispatch (v5)
-// -------------------------------------------------------------------------
-
-#[cfg(target_os = "linux")]
-unsafe extern "C" fn host_gpu_lim_acquire_storage_buffer(
-    handle: *const c_void,
-    byte_size: u64,
-    out_buffer: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_acquire_storage_buffer",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                write_err(
-                    "acquire_storage_buffer: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if out_buffer.is_null() {
-                write_err(
-                    "acquire_storage_buffer: null out_buffer",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            match gpu.acquire_storage_buffer(byte_size) {
-                Ok(buf) => {
-                    unsafe {
-                        std::ptr::write(
-                            out_buffer as *mut crate::core::rhi::StorageBuffer,
-                            buf,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-#[cfg(target_os = "linux")]
-unsafe extern "C" fn host_gpu_lim_acquire_uniform_buffer(
-    handle: *const c_void,
-    byte_size: u64,
-    out_buffer: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_acquire_uniform_buffer",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                write_err(
-                    "acquire_uniform_buffer: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if out_buffer.is_null() {
-                write_err(
-                    "acquire_uniform_buffer: null out_buffer",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            match gpu.acquire_uniform_buffer(byte_size) {
-                Ok(buf) => {
-                    unsafe {
-                        std::ptr::write(
-                            out_buffer as *mut crate::core::rhi::UniformBuffer,
-                            buf,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-#[cfg(target_os = "linux")]
-unsafe extern "C" fn host_gpu_lim_acquire_vertex_buffer(
-    handle: *const c_void,
-    byte_size: u64,
-    out_buffer: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_acquire_vertex_buffer",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                write_err(
-                    "acquire_vertex_buffer: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if out_buffer.is_null() {
-                write_err(
-                    "acquire_vertex_buffer: null out_buffer",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            match gpu.acquire_vertex_buffer(byte_size) {
-                Ok(buf) => {
-                    unsafe {
-                        std::ptr::write(
-                            out_buffer as *mut crate::core::rhi::VertexBuffer,
-                            buf,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-#[cfg(target_os = "linux")]
-unsafe extern "C" fn host_gpu_lim_acquire_index_buffer(
-    handle: *const c_void,
-    byte_size: u64,
-    out_buffer: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_acquire_index_buffer",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                write_err(
-                    "acquire_index_buffer: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if out_buffer.is_null() {
-                write_err(
-                    "acquire_index_buffer: null out_buffer",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            match gpu.acquire_index_buffer(byte_size) {
-                Ok(buf) => {
-                    unsafe {
-                        std::ptr::write(
-                            out_buffer as *mut crate::core::rhi::IndexBuffer,
-                            buf,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
-#[cfg(not(target_os = "linux"))]
-unsafe extern "C" fn host_gpu_lim_acquire_storage_buffer(
-    _handle: *const c_void,
-    _byte_size: u64,
-    _out_buffer: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    write_err(
-        "acquire_storage_buffer: StorageBuffer is not available on this platform",
-        err_buf,
-        err_buf_cap,
-        err_len,
-    );
-    1
-}
-
-#[cfg(not(target_os = "linux"))]
-unsafe extern "C" fn host_gpu_lim_acquire_uniform_buffer(
-    _handle: *const c_void,
-    _byte_size: u64,
-    _out_buffer: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    write_err(
-        "acquire_uniform_buffer: UniformBuffer is not available on this platform",
-        err_buf,
-        err_buf_cap,
-        err_len,
-    );
-    1
-}
-
-#[cfg(not(target_os = "linux"))]
-unsafe extern "C" fn host_gpu_lim_acquire_vertex_buffer(
-    _handle: *const c_void,
-    _byte_size: u64,
-    _out_buffer: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    write_err(
-        "acquire_vertex_buffer: VertexBuffer is not available on this platform",
-        err_buf,
-        err_buf_cap,
-        err_len,
-    );
-    1
-}
-
-#[cfg(not(target_os = "linux"))]
-unsafe extern "C" fn host_gpu_lim_acquire_index_buffer(
-    _handle: *const c_void,
-    _byte_size: u64,
-    _out_buffer: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    write_err(
-        "acquire_index_buffer: IndexBuffer is not available on this platform",
-        err_buf,
-        err_buf_cap,
-        err_len,
-    );
-    1
-}
-
-// -------------------------------------------------------------------------
-// TextureRegistration Arc-handle lifecycle
-// -------------------------------------------------------------------------
-
-unsafe extern "C" fn host_gpu_lim_clone_texture_registration(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_clone_texture_registration",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: `handle` is `Arc::into_raw(Arc<TextureRegistrationInner>)`-shaped.
-            unsafe {
-                Arc::increment_strong_count(
-                    handle as *const crate::core::context::texture_registration::TextureRegistrationInner,
-                );
-            }
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_drop_texture_registration(handle: *const c_void) {
-    run_host_extern_c(
-        "host_gpu_lim_drop_texture_registration",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            // SAFETY: matched with the `Arc::into_raw` in
-            // `TextureRegistration::from_arc_into_raw`.
-            unsafe {
-                Arc::decrement_strong_count(
-                    handle as *const crate::core::context::texture_registration::TextureRegistrationInner,
-                );
-            }
-        },
-        (),
-    )
-}
-
-// -------------------------------------------------------------------------
-// TextureRegistration method dispatch (v6)
-// -------------------------------------------------------------------------
-
-unsafe extern "C" fn host_gpu_lim_texture_registration_texture(
-    handle: *const c_void,
-) -> *const c_void {
-    run_host_extern_c(
-        "host_gpu_lim_texture_registration_texture",
-        || {
-            if handle.is_null() {
-                return std::ptr::null();
-            }
-            // SAFETY: `handle` is `Arc::into_raw(Arc<TextureRegistrationInner>)`-shaped;
-            // the Arc's strong count keeps the inner alive. We return
-            // a pointer to the inner's `texture` field; the caller
-            // (cdylib) deref's it as `*const Texture`. The pointer is
-            // alive as long as the caller's `TextureRegistration` is.
-            unsafe {
-                let inner = &*(handle
-                    as *const crate::core::context::texture_registration::TextureRegistrationInner);
-                &inner.texture as *const crate::core::rhi::Texture as *const c_void
-            }
-        },
-        std::ptr::null(),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_texture_registration_current_layout(
-    handle: *const c_void,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_texture_registration_current_layout",
-        || {
-            if handle.is_null() {
-                return 0; // VK_IMAGE_LAYOUT_UNDEFINED
-            }
-            #[cfg(target_os = "linux")]
-            {
-                // SAFETY: `handle` is `Arc::into_raw(...)`-shaped.
-                unsafe {
-                    let inner = &*(handle
-                        as *const crate::core::context::texture_registration::TextureRegistrationInner);
-                    inner
-                        .current_layout
-                        .load(std::sync::atomic::Ordering::Acquire)
-                }
-            }
-            #[cfg(not(target_os = "linux"))]
-            {
-                let _ = handle;
-                0
-            }
-        },
-        0,
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_texture_registration_update_layout(
-    handle: *const c_void,
-    layout_raw: i32,
-) {
-    run_host_extern_c(
-        "host_gpu_lim_texture_registration_update_layout",
-        || {
-            if handle.is_null() {
-                return;
-            }
-            #[cfg(target_os = "linux")]
-            {
-                // SAFETY: same shape as
-                // `host_gpu_lim_texture_registration_current_layout`.
-                unsafe {
-                    let inner = &*(handle
-                        as *const crate::core::context::texture_registration::TextureRegistrationInner);
-                    inner
-                        .current_layout
-                        .store(layout_raw, std::sync::atomic::Ordering::Release);
-                }
-            }
-            #[cfg(not(target_os = "linux"))]
-            {
-                let _ = (handle, layout_raw);
-            }
-        },
-        (),
-    )
-}
-
-unsafe extern "C" fn host_gpu_lim_resolve_texture_registration_by_surface_id(
-    handle: *const c_void,
-    surface_id_ptr: *const u8,
-    surface_id_len: usize,
-    has_layout: i32,
-    layout_raw: i32,
-    width: u32,
-    height: u32,
-    out_registration: *mut c_void,
-    err_buf: *mut u8,
-    err_buf_cap: usize,
-    err_len: *mut usize,
-) -> i32 {
-    run_host_extern_c(
-        "host_gpu_lim_resolve_texture_registration_by_surface_id",
-        || -> i32 {
-            let Some(gpu) = (unsafe { handle_as_gpu_context(handle) }) else {
-                write_err(
-                    "resolve_texture_registration_by_surface_id: null gpu handle",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            };
-            if out_registration.is_null() {
-                write_err(
-                    "resolve_texture_registration_by_surface_id: null out_registration",
-                    err_buf,
-                    err_buf_cap,
-                    err_len,
-                );
-                return 1;
-            }
-            let id_bytes = unsafe { slice_from_raw(surface_id_ptr, surface_id_len) };
-            let id_str = match std::str::from_utf8(id_bytes) {
-                Ok(s) => s,
-                Err(_) => {
-                    write_err(
-                        "resolve_texture_registration_by_surface_id: surface_id not valid UTF-8",
-                        err_buf,
-                        err_buf_cap,
-                        err_len,
-                    );
-                    return 1;
-                }
-            };
-            let texture_layout = if has_layout != 0 {
-                Some(layout_raw)
-            } else {
-                None
-            };
-            match gpu.resolve_texture_registration_by_surface_id(id_str, texture_layout, width, height) {
-                Ok(reg) => {
-                    // SAFETY: out_registration points at caller-allocated
-                    // stack storage for a `TextureRegistration` value.
-                    unsafe {
-                        std::ptr::write(
-                            out_registration
-                                as *mut crate::core::context::TextureRegistration,
-                            reg,
-                        );
-                    }
-                    0
-                }
-                Err(e) => {
-                    write_err(&format!("{}", e), err_buf, err_buf_cap, err_len);
-                    1
-                }
-            }
-        },
-        1,
-    )
-}
-
 // -------------------------------------------------------------------------
 // RhiCommandQueue Arc-handle lifecycle + create_command_buffer
 // -------------------------------------------------------------------------
